rolf/hydra_plugins/rolf_hydra_plugin/rolf_searchpath_plugin.py
from hydra.core.config_search_path import ConfigSearchPath
from hydra.plugins.search_path_plugin import SearchPathPlugin


class RoLFSearchPathPlugin(SearchPathPlugin):
    def manipulate_search_path(self, search_path: ConfigSearchPath) -> None:
        search_path.append("rolf", "pkg://rolf/config")
